+++ src/intro.js
@@ -1,2 +1,1 @@
-( function ( global ) {
-"use strict";
+( global ) => {