+++ src/node.js
@@ -1,28 +1,6 @@
-/**
- * Node
- *
- * @param {Mixed} value [Optional] Node value
- * @constructor
- */
-function Node ( value ) {
-	this.value = value;
-	this.edges = {};
-}
-
-/**
- * Setting constructor loop
- *
- * @type {Object}
- */
-Node.prototype.constructor = Node;
-
-/**
- * Node factory
- *
- * @method node
- * @param {Mixed} value [Optional] Node value
- * @return {Object} Node
- */
-function node ( value ) {
-	return new Node( value );
+class Node {
+	constructor ( value ) {
+		this.value = value;
+		this.edges = {};
+	}
 }