+++ src/outro.js
@@ -1,13 +1,24 @@
-// Node, AMD & window supported
+let defined = ( arg ) => {
+	return arg !== undefined;
+};
+
+let graph = () => {
+	return new Graph();
+};
+
+let node = ( value ) => {
+	return new Node( value );
+};
+
 if ( typeof exports != "undefined" ) {
 	module.exports = graph;
 }
 else if ( typeof define == "function" ) {
-	define( function () {
+	define( () => {
 		return graph;
 	} );
 }
 else {
 	global.graph = graph;
 }
-} )( this );
+}( this );